Slightly faster getSlot
Summary:
`getSlot` function is a hotspot, and this micro-optimization seems to help a little.

Before this diff the switch was:

```
switch (mask) {
 case 0x00_00_00_00: ...
 case 0x01_00_00_00: ...
 case 0x02_00_00_00: ...
 ...
}
```

Now:

```
switch (kind) {
  case 0: ...
  case 1: ...
  case 2: ...
}
```

JVM is supposed to generate identical machine code before and after this diff, but it either doesn't or maybe it enables it only in C2 compiler.

Reviewed By: aniketmathur

fbshipit-source-id: 31c20fc9d1ab5f1966da48a3b957f0a9c87bf6f9