Include ExternalAction framework classes into buck cli. Get rid of external actions binary
Summary:
This change allow to get rid of "external_actions" resource set by python wrapper, so we don't need to read "buck.external_actions" env variable and pass a path to external_actions.jar to launch external actions. We just add external_actions classes as dependency to buck cli classes and then launch external actions with `ClassLoaderBootstrapper` and pass main class for external actions as parameter to it.

Switching to `ClassLoaderBootstrapper` is an appropriate way to run java process as it would use a separate classloader so that annotation processors do not have their classpaths polluted with Buck's dependencies .

Reviewed By: bobyangyf

fbshipit-source-id: d0bc198db779b9b046dd7d196b8f1479c1565e35