check for project_generator_swift_labels when adding SWIFT_INCLUDE_PATHS
Summary: When generating the `SWIFT_INCLUDE_PATHS` config we need to know if a dependency contains Swift code or not. The existing code checks a libraries inputs, but when the dependency is a codegen library it will have no inputs. This diff adds a new config  `apple.project_generator_swift_labels` to specify labels to look for in a dependency to know if it has generated Swift code.

Reviewed By: milend

fbshipit-source-id: 936b0227afdca86b77452cc95f888f8cb637bb78