Init filesystem after watchman
Summary:
There's unfortunate cyclic dependency between buckconfig, filesystem, watchman and eden filesystem delegate.

Before this diff, initialization worked this way:

* init filesystem
* init buckconfig
* init watchman
* init eden delegate
* patch filesystem with eden delegate

This diff moves filesystem initialization past watchman, so the following diff could do the following:

* init buckconfig
* init watchman
* init eden delegate
* init filesystem
* patch buckconfig with filesystem

Eden initialization is conditional (depends on eden availability), so it is harder to debug, which leads to problems like the one described in this revert: D28210412.

Buckconfig patching with filesystem is harder to miss.

Reviewed By: bobyangyf

fbshipit-source-id: c7bf00a8f240fe88f6b78967d37b3d8ea00a2746