Bytecode intermediate representation
Summary:
Starlark has bytecode, instructions like CP, CALL etc.

Some instructions have hardcoded:
* jump addresses (relative the the beginning of the bytecode)
* constant and object indices
* fixed slot numbers

Bytecode is optimized for evaluation, but not optimized for compilation. For example, operations like:

* compile two expressions independently and concatenate them
* compile expression twice with different option and pick one which works better

is hard with bytecode.

This diff introduces bytecode intermediate representation. Bytecode IR is similar to bytecode with these differences:

* instructions are represented by java objects
* instruction object and const parameters are stored as objects, not as indices in constant and object pools
* local slot numbers are allocated only at serialization
* jump addresses are represented by objects, not as integer offsets, so jump addresses are computed at serialization to real bytecode
* bytecode is position-independent: for example, concatenating two IR produces valid IR

This change makes the compiler safer and easier to patch. Previously the compiler had saveState/reset operations for bytecode writer, which is error prone and not flexible enough.

For example, with saveStack/reset it is not possible to implement this pattern:

* compile expression x
* compile expression y
* discard x and compile it again with different options

It is done mostly for future changes in the interpreter, but for now for example, new IR is used to make `if` statement compilation more efficient: previously we compiled `if` condition twice: first time to determine if it is a constant, and second time again when actually generating bytecode. Now if condition is only compiled once. Or call compilation is much more clean (and safe) now.

Reviewed By: aniketmathur

fbshipit-source-id: bb2bb3ebb456a080afa3328964ff7d76fa5a422e