Pass Watchman to ProjectFilesystem construction
Summary:
`EdenProjectFilesystemDelegate` needs watchman.

Previously it was initialized after the construction, which was hard maintain (track all paths where we have to set the watchman).

Now watchman is required to construct `ProjectFilesystem`.

This is done for the following diff D28244006, which was reverted because watchman in eden was set twice when fs is eden and `--reuse-current-config` option was used. This situation is not possible now.

Reviewed By: bobyangyf

fbshipit-source-id: 73cd42d7c99ab3232a9181dbd2d165b7e71e5a01