allow named variants
Summary: In order to use result of genrule as a variant in `apple_resource` we need to be able to specify the localization for this variant. Currently the localization is inferred from the path, basically from the directory name resouce file is residing in. It's a frequent pattern to add `/LOCALE.lproj` suffix to genrule name in order for such directory to be created when genrule output is used as a variant. Though it's wrong to assume how future versions of Buck are going to name intermediate directories. As a more robust approach I'm introducing `named_variants` attribute to `apple_resource`, it's a dictionary where keys are locales and values are lists of variant files for specific locale (without the constraint for them be be contained in a directory ending with .lproj).

Reviewed By: milend

fbshipit-source-id: 8f0a6da732892b614d8b944248aa74adce78feb2