Implement custom LinkedHashMap and fix linking bug
Summary:
This diff improves performance and fixes a bug.

# The bug

The bug is this:

```
def foo(a, **kwargs): pass

foo(a=1, **dict(a=2))
```

This invocation should produce an error, but before this diff this invocation bound first `a` to local `a` and put second `a` to `kwargs`. Which is incorrect.

Simply fixing this bug with `Dict` implemented with `java.util.LinkedHashMap` implementation would cause measurable regression probably 2% (I did not try, I discovered this bug while implementing this optimization).

Note interpreter spends 1/7 of total time processing function arguments, so even small changes in function invocation handling have significant effect.

# DictMap

This diff implements custom `LinkedHashMap` (named **`DictMap`**) which is structurally similar `java.util.LinkedHashMap` with following properties:
* implementation is simpler, there's no RB tree in it, which is useful in certain situations, but in Starlark it only causes overhead (also, it is hard to implement)
* exposes internals

Implementation of `LinkedHashMap` (without RB tree part and without boilerplate) is just 10 lines of code or so, not a rocket science.

It is important that `DictMap` exposes internals for several reasons. I'll show several examples:

## `putNoEvict`

Consider a function:

```
def foo(**kwargs)
```

and invocation:

```
foo(a=1, b=2)
```

So we need:
* create a map
* put `a=1` into the map
* put `b=2` into the map

When insering `b=2` into map, we know that the key is unique, but `LinkedHashMap` does not know that, and for each inserted key it queries it first. That makes `put` operation much more expensive than `putNoEvict`.

Because `get` (and accordingly `put`) is branchy with all bad consequences for the CPU. But `putNoEvict` (especially `putNoEvictNoResize`) does virtually no branches.

## reuse hashes from dict arg

Consider a function:

```
def foo(**kwargs): ...
```

and invocation:

```
foo(**kwargs)
```

We need to invoke `new LinkedHashMap(oldMap)`.

Default java `LinkedHashMap` implementation simply iterates over `oldMap` and for each key it recomputes hash codes. Although caches are usually stored in objects (e. g. `String` caches hash code), computing caches is still not free.

But when we computing `new DictMap(oldMap)` we only accept `oldMap` of type `DictMap` and we reuse the hash codes which are stored in hashmap nodes. Which makes generated machine code much more efficient.

# precompute caches for named arg calls

Consider a function:

```
def foo(**kwargs): ...
```

and invocation:

```
foo(a=1, b=2)
```

If we used `LinkedHashMap`, we would have add `a=1` and `b=2` to that `LinkedHashMap`, and `LinkedHashMap` would have to compute the hashes each time this function is invoked. Which is relatively cheap (because hash is cached in `String`), but not when the function is invoked a lot.

But our new `DictMap` implementation has a `put` variant which accepts `keyHash` param, so there's no hashcode computation when function is invoked.

# query nodes, not just values

Consider this function:

```
def foo(b, **kwargs_there):
```

and invocation:

```
foo(**kwargs_here)
```

When processing parameters (like `b`), we store parameters names which were used from `kwargs_here`.

So when we reconstructing dict (`kwargs_there`), we iterating over `kwargs_here` and skipping previously processed parameters.

Name `b` and key `"b"` in `kwargs_here` are equal, but not necessary the same java objects, so we had to use equality.

`DictMap` exposes `getNode` function which returns actual hashtable node. So when we iterating over kwargs argument `kwargs_here`, we can use identity comparison of nodes instead of equality comparison of keys. Which is much faster.

`LinkedHashMap` does not provide access to nodes.

# Also `BuiltinFunctionLinkedError`

This diff simplifies/speeds up implementation of `BuiltinFunction` similarly how it is done for `StarlarkFunction`: there's a separate class `BuiltinFunctionLinkedError` (similar to `StarlarkFunctionLinkedError`) which is optimized for readability and producing correct errors. `BuiltinFunction` now delegates to `BuiltinFunctionLinkedError` on any error.

Reviewed By: mzlee

fbshipit-source-id: 6972b00779cc659f12196fbeae6e4cdfda82f57f