Do not compile unreachable statements
Summary:
```
def foo():
  return 10
  print(20) # this is dead code, no need to compile it
```

This has no measurable performance effect, but better land to avoid overhead if compilation becomes expensive.

Note unreachable statements do not occur naturally in code, but may appear when optimizer runs, for example:

```
def foo():
  if native.read_config("foo", "bar"):
    return
  a
  lot
  more
  code
```

`native.read_config` is inlined at compile time, so when it is inlined as true, `if` statement is replaced with `return`, and the following code is unreachable.

Reviewed By: ndmitchell

fbshipit-source-id: 18bbfb2e83defa556dac95b51e79daa725fc9be8