Remove disableWithFlag and enableOnlyWithFlag
Summary:
Starlark has (had) builtin function annotation for functions and parameters like `disableWithFlag` and `enableOnlyWithFlag` which enabled certain functions and parameters depending on semantics.

We don't use that feature, and it creates some runtime overhead. If we ever need it, we can reimplement them without runtime overhead.

Annotations are removed from `StarlarkMethod.java` and friends, and the rest of the diff is remove of everything which depends on these annotations.

Reviewed By: aniketmathur

fbshipit-source-id: 150961fd915bc54985fd30396a13fc46ed8cde3a