javacd: fix for chrome trace events
Summary:
This is a temporary fix that allows to have valid chrome traces if you are running with javacd.

The issue is the following:
T1. Buck starts pipelining rule for 2 commands (abi.jar + library.jar)
T2. Buck prepares both commands into 1 protobuf message and sends this message to javacd process. Buck creates a map for result, where keys are action ids for commands and values are futures that wait for ResultEvents with an appropriate action id from javacd.
T3. Javacd reads received message and start execution the first command ( = abi.jar)
T4. When abi.jar is ready, javacd prepares ResultEvent with action id for abi.jar command and sends it back to Buck.
T5. Buck reads ResultEvent for abi.jar and marks ResultEventFuture with the result.
T6. Buck's step associated with abi.jar finished with successful result.

T7.
* Javacd started execution of library.jar command received at T3. (Happens somewhere after T4 till T7)
*  At the same time Buck's pipelining runner starts a new step related to library.jar command

T8. When library.jar is ready, javacd prepares ResultEvent with action id for library.jar command and sends it back to Buck.
T9. Buck reads ResultEvent for library.jar and mark ResultEventFuture with the result.

**The issue**: `T7`, as buck's start event for library.jar step (happened in the buck process) occurred after events from library.jar execution in the javacd occurred and processed.

**Chrome trace viewer requires events inside the same thread to be ordered by time.**

**The fix in this diff**: adds delay in T7 before starting execution of the received library.jar command

**Next steps**
I plan to remove this temporary delay and change the communication in a way that javacd would be notified by buck when it could start execution of library.jar command after abi.jar command is ready. This change requires to extend commands list with a new command (ex: StartNextCommand ), to send a new command from buck to javacd.

Reviewed By: bobyangyf

fbshipit-source-id: 8f61488efeddd510c81c74af17771486438d2ae0