simplify alloc/release in bytecode IR
Summary:
Java Starlark is a register machine.

Registers are locals variables or temporaries.

Temporary variables need to be allocated before the first instruction which references it and release it after the last instruction which references it, so the same register can be used by the following instructions.

Previous diff D28278628 implements too complicated and error-prone algorithm: there's special (compile time, IR) `AllocSlot` instruction, each temporary (called `LazyLocal`) has a reference counter, and each time that temporary is referenced by an instruction at serialization, that refcounter is decremented. And temporary is deallocated when refcounter reaches zero.

The problem with that approach is that refcounter needs to managed manually (with `incRef` and `decRef` functions).

New implementation removes `AllocSlot` instruction, now the scope of each temporary is determined at the beginning of the serialization by inspecting all instructions.

The algorithm is quite simple, it lives in `BcIrWriteContext`.

With new implementation, all bytecode instructions are immutable.

The compilation becomes slightly more expensive, but it is likely not significant.

Reviewed By: ndmitchell

fbshipit-source-id: 9fdffa4a802e1ba6f6ee6cb510191a0ff4ed871c