validation token
Summary:
Invalidation concurrency model is this.

We have:

* invalidation counter
* read-write lock

When invalidating the cache, we:
* acquire the lock for writing
* increment invalidation counter

When starting command, we acquire the invalidation counter (it is called validation token).

When updating the cache, we:
* acquire the lock for reading
* check that validation token is still valid
* update the cache only id validation token is still valid

When reading from the cache, we:
* speculatively fetch data from cache
* check the validation token *after* data is fetched from cache
* if validation token is not valid, discard the fetched data

We must acquire read lock for writing to the cache because cache update is multi-step operation which cannot be performed atomically with checking validation token.

We can avoid acquiring lock for reading from the cache because reading from cache has no side effects.

This diff finishes the implementation of invalidation from [the proposal from the Workplace](https://fburl.com/9l2vzqt5).

Reviewed By: bobyangyf

fbshipit-source-id: 1f36e4e2c0bb23325bb1e43631e0717c8d888234