javacd: Pipelined rules call appendWithCommonSetupSteps() for all stages
Summary:
We have to call `appendWithCommonSetupSteps` not only for the first stage, but for every other stage in the pipeline as it would create missing directories and record outputs as result.

These calls were missed in the original implementation and causing issues ( default output directories were not created / artifacts are not included in the cache) if execution has been performed with javacd (nothing was recorded, nothing was expected :) ).

Reviewed By: bobyangyf

fbshipit-source-id: 947cb826f4db560fcf327245a99dbe5512f0a3c3