JavaCD steps builders invoke buildableContext.recordArtifact to declare their outputs.
Summary:
If build with javacd is enabled then `JavaCDAbiJarStepsBuilder` and `JavaCDLibraryJarStepsBuilder` are used for creating steps that buck would  execute. Only 1 step `JavaCDWorkerToolStep` would be created. This step would communicate with javacd and eventually notify buck that the result is ready.

What we are missing here is a way to verify that output is excepted to the same that buck calculates by reflection from MBR rule. To archive that I have added `recordArtifactVerifier.recordArtifact` to verify an output of the rule.

Reviewed By: bobyangyf

fbshipit-source-id: c78b759ef374984cd4d235ba3e0683621bbb7e34