Move allowed argument error message into BuiltinFunctionLinkedError
Summary:
Previously generated code was responsible for producing error message when parameter type is incorrect (e. g. is a parameter is declared to accept only `String`, but argument was `None`).

Now generated code still validates the argument, but instead of producing nice error message, generated code just throws `ArgumentBindException`, and nice error message is produced by `BuiltinFunctionLinkedError`, which handles `ArgumentBindException`.

Note `BuiltinFunctionLinkedError` is a special version of argument binding for `BuiltinFunction` which is optimized for readability/simplicity, not for performance.

This change simplifies generated code (note the changes in `MethodDescriptorGen.java`), and could actually speed-up it.

Generated code for `MethodLibrary` now looks like this: P415043097.

Reviewed By: ndmitchell

fbshipit-source-id: 17dfa3b6e0d231990984c473c1e928959f1e5094