de-duplicate compiler flags in xcconfig files
Summary:
This diff changes the generated xcconfig settings for `OTHER_C*FLAGS`, `OTHER_LDFLAGS` and `OTHER_SWIFT_FLAGS` to be override settings, not appended settings. The output change is that the configs will not have an `$(inherited)` prefix. This was causing flag duplication as we ended up with settings of the form:

```
OTHER_CFLAGS = $(inherited) ...
OTHER_CFLAGS[sdk=iphonesimulator*][arch=arm64] = $(inherited) ...
```
which means that when matching one of the SDK and architecture specific configs we would inherit from the base setting. There is no reason to inherit for these settings, we fully specify the flags for each value.

Reviewed By: williamtwilson

fbshipit-source-id: 2377d2f15ffbb9bcfa791b78da25646eb383ecb2