add apple.project_generator_index_via_build_flags
Summary:
This diff adds a new project generation option `apple.project_generator_index_via_build_flags`. When true the project generator will do a single pass of the target to gather Buck output paths to use as include search paths. This avoids the necessity of generating separate header maps from Buck and should more closely match the indexing commands to the Buck compile commands.

The required Buck targets are not currently built as part of `buck project`, that will be added in a follow up diff.

Reviewed By: milend

fbshipit-source-id: 5ec31862c62a91553003a8ae1c71a3ef478da161