Make input rule key size limit per-input
Summary:
This changes the semantics of the input-based rule key input
size limit to make the limit apply to each input to the rule key, rather
then the sum of all inputs.  When it was originally added, a while ago,
it was meant to address two issues:
1) Large static links with lots of inputs were taking a long time to (at the
    time serially) calculate the hashes of all their inputs.
2) Some genrules would process really large statically linked inputs,
    which would take a long time to hash.

Now that Buck hashes build rule outputs as it builds them and caches
them, (1) isn't really an issue anymore (and, in fbcode, we completely
disable the input rule key limit for `dev` mode).  So, this diff changes the
input rule key limit to only apply to each individual input, so that we can
enable input rule keys in more scenarios (e.g. T91563697).

Reviewed By: mzlee

fbshipit-source-id: 8c0f2f91e918cf85881271fabe0ebb3195336de5