DefaultIsolatedEventBus writes default action id in case other is not specified
Summary:
Protobuf doesn't allow to set `null` value to a field, so need to specify default value for actual id in `DefaultIsolatedEventBus`.

`DefaultIsolatedEventBus` is used in external actions framework and javacd to send events back to buck. Some steps scheduled to execute with external actions could not specify action id (ex. PerfEvents.scope), so need to set a default value for action id.

Reviewed By: stepancheg

fbshipit-source-id: f818990bb975f6bc78199865619225f12a1caf19