an option to disable floats
Summary:
When semantic option is enabled, floats are almost unavailable.

There's a couple of places where floats still can be used, for example, in dead code:

```
def f():
  return 1.0
```

this code compiles fine, but fails at runtime. Not perfect, but OK for the task of getting rid of floats in repositories.

This diff only adds an option to Starlark, following diff enables this option in Buck.

Reviewed By: ndmitchell

fbshipit-source-id: 3e528df75d9ef7b2b37833164179463883072df5