Put tremendous preprocessor flags into argsfile
Summary:
We ran into the ARG_MAX limit again for haddock when attemping to build
haddock and hoogle database for all Haskell libraries in facebook. Here are the
arguemnts

```
$ sort | uniq -c | sort
     12   -idirafter
     91   -isystem
    381   -expose-package
    755   --read-interface
    757   -package-db
    962   -hide-package
   6117   -I
  12510   -optP
  29428   --optghc
```

preprocessor flags are the majority ones, let's put them in argsfile as well.
We attempted to handle flags like `-optP=foo` in wrapper, but not `-optP -I -optP path`,
so this should be ok.

Reviewed By: pepeiborra

fbshipit-source-id: 16ddc157090a6b852064c6cdd4c0844d5db6e3eb