WorkerProcessPool: replace WorkerLifecycle array with an ImmutableList of DefaultWorkerLifecycle
Summary:
During the `WorkerProcessPool `creation we store an array of `WorkerLifecycle` and use it only in the `close()` method.

No need to maintain an array of `WorkerLifecycle` with size of `maxWorkers * maxInstances`. It could be replaced with an `ImmutableList` with only `maxWorkers` size and later we would invoke `close()` on these elements.

Changes:
* Store `ImmutableList` of `DefaultWorkerLifecycle`.
* `WorkerLifecycle` no longer implements `Closeable`
* `DefaultWorkerLifecycle` implements `Closeable`
* `DefaultWorkerLifecycle` verifies that `close()` is called only once.
* Updated unit tests with maxInstance param

Reviewed By: bobyangyf

fbshipit-source-id: 36124bdec1a8660d4a3bbe34eed4ae037ac01212