don't override AppleToolchain strip args
Summary: This change will not overwrite the strip args defined in an apple_toolchain. This allows the toolchain to define its strip args specific to the strip tool it is using. The `apple.strip_swift_symbols` config will no longer apply to the strip args of an apple_toolchain, it is not possible to map an arbitrary set of strip flags to a Swift strippable set, and this can be set directly in the toolchain target instead.

Reviewed By: blackm00n

fbshipit-source-id: 43d504d06c6e307b927e4e9f313fbfded96e8eee