Fix SymlinkCache
Summary:
Symlink cache contained a map by relative path. Now it contains a map by (cell, relative path).

If two cells contain the same path, and one path is symlinked, there was a race.

This is clowntown, I spent a day debugging this when I switched Buck to Starlark by default (following diff D28986506), because tests started failing only on sandcastle, and it was not reproducible locally.

The problem was revealed when Starlark was enabled by default because of different speed of evaluation of Python vs Starlark.

Reviewed By: mykola-semko

fbshipit-source-id: 6f311b87f37cb9bce2ce066e9f758dbad4dbe144