Switch default syntax to Starlark
Summary:
When `parser.default_build_file_syntax` unspecified, use Starlark. Previously Python DSL was used.

To make sure we don't have projects which implicitly rely on Python DSL.

This is single-line change in `src` and the rest of the diff is test fixes.

Some tests are disabled because certain features (like byte counting) are not implemented in Starlark parser. These can be switched to run Python DSL explicitly, but the next step is removal of Python DSL parser anyway.

Reviewed By: bobyangyf

fbshipit-source-id: 683c5dd47031481d7ff19d69453f91cec904ef0c