Remove windows-specific path relativization from HybridGlobberFactory
Summary:
This is some strange code.

The code was added in D12944598, see the discussion in the [linked thread](https://fb.workplace.com/groups/askbuck/permalink/2168362386545698/). Function was added by someone who is no longer works with us.

This code is not used now after previous refactorings (relativization is not needed anymore).

I'm not sure what was the original problem, perhaps absolute either path passed to the function or hybrid globber root was not normalized properly.

CC people who participated in the discussion of the original issue.

Reviewed By: akrieger

fbshipit-source-id: 6c6e1b9843aaa9c3c72b5f29093d7b08b8d2529d