Add the native library map to rulekey
Summary:
Prior to this diff if you were to create a shared module that only
contained a native library - the rulekey for the android_modularity
rule would be unchanged.

This change makes it so that that edge case is no longer true.

Note:
It is still possible that modifying the merge map will produce an
identical rule key this will need to be addressed in a further diff.

Reviewed By: christolliday

fbshipit-source-id: 8735f156d56fb573097431d819f9df358d4a261f