assert we construct WatchmanWatcher with the same keys in two parameters
Summary:
If that assertion does not hold, I'm not sure `WatchmanWatcher` works correctly.

And if assertion does not hold, something will break, and we will document how it is possible.

I think this assertion is true since D27179706, where we either run with watchman for all cells, or do not run watchman at all.

Reviewed By: mykola-semko

fbshipit-source-id: d4a63b5e09a7c0e91acc9c745960415994ab74e4