Format "N MORE THREADS" to be constant length
Summary:
Currently, when running `buck build` results in too many threads to fit on one screen, additional threads are shown in a collapsed view at the bottom of the execution list, as:
```
 - N MORE THREADS: [:] [:] [:] [:]
```

Because `N` is just an integer appended to the string, when the number of additional threads changes from single- to double-digits, the rest of the line shifts by a character:
```
 - 10 MORE THREADS: ['] ['] ['] ['] ['] ['] ['] ['] ['] [']
 - 9 MORE THREADS: [:] [:] [:] [:] [:] [:] [:] [:] [:]
 - 12 MORE THREADS: [.] [.] [.] [.] [.] [.] [.] [.] [.] [.] [.] [.]
```
This is distracting to see out of the corner of one's eye, and also makes it a little harder to actually make sense of the status visually in cases when the thread count is often flipping between the two.

This diff formats the number of extra threads to a standard width. This way there is less change when the number of threads changes:
```
 -  10 MORE THREADS: ['] ['] ['] ['] ['] ['] ['] ['] ['] [']
 -   9 MORE THREADS: [:] [:] [:] [:] [:] [:] [:] [:] [:]
 -  12 MORE THREADS: [.] [.] [.] [.] [.] [.] [.] [.] [.] [.] [.] [.]
```

Reviewed By: mykola-semko

fbshipit-source-id: 48b3f9b5acbd7a500c1af408a6f5c283d56f9f36