Store Location stack per instruction
Summary:
Before this diff each instruction contained a location.

Now each instruction contains a location stack.

This is a preparation to implement function inlining (D28611893): when function is inlined, and inlined function fails, full stack trace need to be preserved as if function was not inlined. So each instruction of inlined function contains two (or more) locations: location of caller and location of inlined instruction.

Most of the diff is mechanical replacement `BcWriter.LocOffset` with `ImmutableList<BcWriter.LocOffset>` (one location with stack).

Reviewed By: bobyangyf

fbshipit-source-id: 0efbc99629d482e191c115ae8d2e8fcb7e29fb64