Remove test-only JsFile.getVirtualPath method
Summary:
Refactors `JsLibraryDescriptionTest` to avoid needing to expose the `getVirtualPath` method on instances of the `JsFile` rule. Instead, we test the same functionality by inspecting the `WorkerShellStep` that the rule generates, where the virtual path is passed as `sourceJsFileName`.

An upcoming diff will move the logic for deriving virtual paths into `JsFile`, as part of adding support for passing additional sources to file rules.

Reviewed By: mykola-semko

fbshipit-source-id: 1c380ce3ca751dfb1749e2be53cf1194f056ee1d