Consolidate JsFile path data in JsSourcePath
Summary:
Creates the `JsSourcePath` class to represent a path to a source file used in a JavaScript build.

The `srcs` argument to `js_library` can hold file paths (e.g. from `glob()`), references to files within directories (as a path + inner path), or a combination of both. Previously we've used a cumbersome `Either<>` type to encode this data throughout `JsLibrary` and `JsFile`. Here we wrap this data in a clearer type and consolidate some other logic around it - particularly the logic for deriving a "virtual path" from a source path, which now lives entirely in `JsFile`.

An upcoming diff will rely on this setup to pass additional `JsSourcePath`s to `JsFile`.

Reviewed By: mykola-semko

fbshipit-source-id: 1976e6879029f672ffa3727a279ac57e9cbdefe4