handle pch includes when apple. project_generator_index_via_build_flags=true
Summary: We need to handle precompiled header targets when generating Xcode project configs when `apple. project_generator_index_via_build_flags=true`. It would be preferred to rely on modules for this case, but we already have code relying on the pch files being present that omit imports of Foundation and UIKit.

Reviewed By: milend

fbshipit-source-id: 27d52b4764c8c70f9c384c0f9b22cf4f8f4315c0