Fix perf event ordering
Summary:
D21830522 added the explicit `close` of the `package_pipeline` event scope but unfortunately did so in a way that messed up its ordering relative to the `raw_target_node_parse_pipeline` event scope. Given that they're logged on the same thread, they're treated as nested scopes so they need to open/close as A->B -> B->A but it's currently A->B -> A->B, which throws off any loader code that validates nesting correctness.

This diff reorders the logging (which happens as part of `.close()`) so that it's correct.

Impact: `chrome://tracing` doesn't seem to care about this but Tracery's importer does.

Reviewed By: milend

fbshipit-source-id: cfb577e3863d57e59708b32e6a61934e247f117c