Prepare js_file & js_library for asset file grouping
Summary:
Structures JsLibrary so it can distribute source files to JsFile rules by canonicalising the source file names. Each JsFile receives one JsSourcePath as the "main source" and zero or more JsSourcePaths as "additional sources".

For ease of review/testing, in this diff, **we use the identity function to canonicalise names**. As a result, each JsSourcePath becomes the main source of a JsFile, with zero additional sources - exactly the same as before additional sources were supported.

In a separate diff we will change the `canonicalizeBaseName` method to actually implement a 1:many mapping of JsFiles to source files, specifically by stripping any platform and resolution information from asset files.

Reviewed By: mykola-semko

fbshipit-source-id: 87aba687a41cac7b35181fe96a7fea3abca791c3