Do not start a daemon if no daemon is running
Summary: Every time the plugin tries to get the http port of the server, it starts a buck daemon. With this change, the behavior is now changed and we no longer start a new daemon if there is no buckd running. This also means we have to check the connection every time we try to show something in the buck toolwindow, while previously we only check the connection during project startup.

Reviewed By: ardavank

fbshipit-source-id: 178c085ef1b0266e792dab861ee25c65fe1bf303