Fix Mach-O UUID Extraction for >= 2GiB
Summary:
`FileChannel.map()` is only able to address `Integer.MAX_VALUE` bytes, but some Mach-O binary files can be larger. When mapping in a Mach-O file to extract its UUID, limit the mapped data range to [0, `Integer.MAX_VALUE`] to prevent `FileChannel` from throwing an exception.

This truncation is acceptable because the logic only needs access to the array of `load_command` derived structs that follow the `mach_header[_64]`. To be safe, though, if there is a wacky Mach-O file that has more than 2GiB of load command data, the UUID extraction function will throw an exception.

Reviewed By: milend

fbshipit-source-id: b77ae585259bb2fdff5e4217e566f8551bb2e545