add `use_swift_driver` to control Swift compilation mode
Summary: This diff is the first step for enabling migration to use the Swift driver. This adds the plumbing from targets to the SwiftCompile steps and removes the previously hardcoded `-frontend` flag injection in AppleCxxPlatforms and SwiftToolchain. Swift compilation with `use_swift_driver = True` will not work yet, the flags will need to be filtered and we will need to specify the output paths using a JSON output map.

Reviewed By: drodriguez

fbshipit-source-id: 2aa7790e4c373ca3d4b64badf474f4d6dfc3a784