Fix FileClassPathRunner when no classpath argfile is specified.
Summary:
So, FileClassPathRunner is used to backdoor a large classpath into java
when running things, like instrumentation tests. On Linux buck seemingly always
generates a classpath argfile, and things work fine. On Windows, it does not, and
things fail to run. On inspection, the condition that FileClassPathRunner was using
to check if it should read the argfile was not entirely accurate. It tested for
the presence of `buck.testrunner_classes`, which was always set, but instead should
have been testing for `buck.classpath_file`. As a result, it was falling into code
to modify the classpath but only appending the value in `buck.testrunner_classes`,
and thus blowing away the actual entries in -classpath that buck specified. This
caused instrumentation tests on Windows to fail with `java.lang.NoClassDefFoundError: com/android/ddmlib/IShellEnabledDevice`
because the path to that jarfile, ddmlib.jar, was now missing.

Fixing the test to check for the argfile itself allows the tests to pass.

Reviewed By: jdonald, bbk79, bobyangyf

fbshipit-source-id: 7589c6ea37444be63e806b104be3d53e0efc8039