use correct flags when use_swift_driver=true
Summary: This change alters the flags used as part of Swift compilation so that targets can opt in to using the Swift driver. The compilation model is fixed to be the same as the frontend: WMO with single object file output. This will have to be more flexible if we want to take advantage of batch mode or cross module incrementality.

Reviewed By: drodriguez

fbshipit-source-id: fa758f306b95a5357ae6e22afefd98064554d0c5