Support python base_module
Summary: Python plugin doesn't support package prefix so we have to make sure we put the module at the right place based on `base_module` property of the rule. For example, for a module path of `a/b/c/d` and `base_module="c.d"`, we should generate the module at `a/b` instead of `a/b/c/d` so that the directory structure of the module (`c/d/`) matches the `base_module` (`c.d`).

Reviewed By: ardavank

fbshipit-source-id: 5128b327fe7c85c817a491f7545bcd743a1250f4