Make strip flags configurable
Summary:
Add `strip_apk_libs_flags` to `ndk_toolchain` to customize the flags passes to the `strip` command, but keep the default flags as `--strip-unneeded`.

We have special MIP builds on fb4a that require some extra flags to be passed to `llvm-strip` so that a metadata section is removed. D30088915 uses this new flag to do this.

fbshipit-source-id: b902cb15a3f3246ad057ec156eb5ea9203eb082b