Refactor Infer/clang flavour to use InferPlatform
Summary:
Main reason for doing this work is to enable RE for Infer/clang.

This diff refactors #infer-capture-all flavour, bringing it into line with the infer-java flavour when it comes to Infer configuration and 'Platform'.

A test was removed (inferCaptureAllIncludesExportedDeps) the use case they were covering is covered by other tests and they were to intrusive to fix with the new configuration setup.

Reviewed By: ngorogiannis

fbshipit-source-id: 853c41c8114b27ebb4e67293990b5971877004e6