Fix the generation of modulemaps for `buck project`.
Summary:
With the changes from D30110026, the relative path used by `buck project` for
the `-Swift.h` header stopped working. For the `buck project` invocation the
`module.modulemap` file still sits side by side with the headers.

The reason is that the `ModuleMap` renderer is the one that decided the path,
instead of being given the actual path that the caller was using. The changes
involve removing the enumeration that used to decide if the Swift module was
needed or not, and passing in an `Optional` with the actual path of the Swift
header (as calculated by the caller).

The call sites of `ModuleMap.create` have been updated to reflect the new API,
as well as the tests.

There's also other improvements in the `ModuleMap.java` file to avoid repeating
the module map rendering code for each of the three cases. Everything should be
possible through the `Module` renderer: several header paths for the single
module mode, and one header path for the submodule mode; printing or not the
`export *` line for user header modules vs. the Swift module; adding different
requirements, like `cplusplus` when required or `objc` for the Swift module.`

fbshipit-source-id: d9f0c43142bf7281a884a89e9dd0760904ea4268